added test option to mfa exec
